+++ opentelemetry-appender-tracing/tests/test_no_stack_overflow_after_shutdown.rs
@@ -1,5 +1,5 @@
 use opentelemetry_appender_tracing::layer;
-use opentelemetry_sdk::logs::SdkLoggerProvider;
+use opentelemetry_sdk::logs::{BatchLogProcessor, SdkLoggerProvider};
 use tracing::info;
 use tracing_subscriber::{layer::SubscriberExt, util::SubscriberInitExt};
 
@@ -7,9 +7,11 @@ use tracing_subscriber::{layer::SubscriberExt, util::SubscriberInitExt};
 fn test_logging_after_shutdown_does_not_cause_telemetry_induced_telemetry() {
     //! Reproduces [#3161](https://github.com/open-telemetry/opentelemetry-rust/issues/3161)
     let exporter = opentelemetry_stdout::LogExporter::default();
+    let processor = BatchLogProcessor::builder(exporter).build().unwrap();
     let provider: SdkLoggerProvider = SdkLoggerProvider::builder()
-        .with_batch_exporter(exporter)
-        .build();
+        .with_log_processor(processor)
+        .build()
+        .unwrap();
 
     let otel_layer = layer::OpenTelemetryTracingBridge::new(&provider);