+++ opentelemetry-sdk/src/logs/log_processor_with_async_runtime.rs
@@ -495,7 +495,8 @@ mod tests {
                     ])
                     .build(),
             )
-            .build();
+            .build()
+            .unwrap();
         assert_eq!(exporter.get_resource().unwrap().into_iter().count(), 5);
     }
 
@@ -519,7 +520,8 @@ mod tests {
                     ])
                     .build(),
             )
-            .build();
+            .build()
+            .unwrap();
 
         provider.force_flush().unwrap();
 
@@ -678,7 +680,8 @@ mod tests {
         let logger_provider = SdkLoggerProvider::builder()
             .with_log_processor(first_processor)
             .with_log_processor(second_processor)
-            .build();
+            .build()
+            .unwrap();
 
         let logger = logger_provider.logger("test-logger");
         let mut log_record = logger.create_log_record();
@@ -786,7 +789,8 @@ mod tests {
                     ])
                     .build(),
             )
-            .build();
+            .build()
+            .unwrap();
         provider.force_flush().unwrap();
         assert_eq!(exporter.get_resource().unwrap().into_iter().count(), 5);
         let _ = provider.shutdown();